Refactor: prepare to refine signature parsing process with nom
